backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
alembic==1.17.2
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
audioop-lts==0.2.2
beautifulsoup4==4.14.2
blinker==1.9.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cloudpathlib==0.23.0
colorama==0.4.6
comtypes==1.4.13
confection==0.1.5
contourpy==1.3.3
cycler==0.12.1
cymem==2.0.11
deep-translator==1.11.4
docopt==0.6.2
edge-tts==7.2.3
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastjsonschema==2.21.2
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.60.0
frozenlist==1.8.0
glfw==2.9.0
greenlet==3.2.4
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huepy==1.2.1
idna==3.11
itsdangerous==2.2.0
jaraco.context==6.0.1
Jinja2==3.1.6
Js2Py==0.74
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_core==5.9.1
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
livereload==2.7.1
Mako==1.3.10
marisa-trie==1.3.1
markdown-it-py==4.0.0
markdown2==2.5.4
MarkupSafe==3.0.3
matplotlib==3.10.6
mdurl==0.1.2
more-itertools==10.8.0
MouseInfo==0.1.3
multidict==6.7.0
murmurhash==1.0.13
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.2.0
mysqlclient==2.2.7
narwhals==2.10.1
nbformat==5.10.4
numpy==2.2.6
opencv-python==4.12.0.88
outcome==1.3.0.post0
packaging==25.0
pillow==11.3.0
pipwin==0.5.2
platformdirs==4.5.0
plotly==6.3.1
preshed==3.0.10
propcache==0.4.1
protobuf==4.21.12
psutil==7.1.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.19.2
pyjsparser==2.7.1
PyMsgBox==2.0.1
PyOpenGL==3.1.9
pyparsing==3.2.5
pyperclip==1.11.0
pypiwin32==223
PyPrind==2.11.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.1
PyRect==0.2.0
PyScreeze==1.0.1
pySmartDL==1.3.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pyttsx3==2.99
pytweening==1.2.0
pywhatkit==5.4
pywin32==311
RapidFuzz==3.14.1
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.28.0
selenium==4.37.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.4.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.14.3
SQLAlchemy==2.0.44
srsly==2.5.1
standard-aifc==3.13.0
standard-chunk==3.13.0
tabulate==0.9.0
thefuzz==0.22.1
thinc==8.3.6
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.31.0
trio-websocket==0.12.2
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
vite==1.5.2
wasabi==1.1.3
weasel==0.4.1
websocket-client==1.9.0
Werkzeug==3.1.3
wikipedia==1.4.0
wolframalpha==5.1.3
wrapt==2.0.0
wsproto==1.2.0
xmltodict==1.0.2
yarl==1.22.0
